voice_message_backend/pyproject.toml
[tool.poetry]
name = "voice-message-backend"
version = "0.1.0"
description = ""
authors = ["nahu02 <nahu1002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
ell-ai = {extras = ["all"], version = "^0.0.15"}
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
openai = "^1.60.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "voice_message_backend.main:main"
dev = "voice_message_backend.main:run_dev"